% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorModels.R
\name{accumulateTrans}
\alias{accumulateTrans}
\title{Sum matrices of transition counts together, accounting for the possibility
of variation in the number of columns present in each.}
\usage{
accumulateTrans(trans)
}
\arguments{
\item{trans}{(Required). A list of matrices recording the counts of transitions in each sample.}
}
\description{
Sum matrices of transition counts together, accounting for the possibility
of variation in the number of columns present in each.
}
