% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{assignTaxonomy}
\alias{assignTaxonomy}
\title{Classifies sequences against reference training set.}
\usage{
assignTaxonomy(seqs, refFasta, minBoot = 50, verbose = FALSE)
}
\arguments{
\item{seqs}{(Required). A character vector of the sequences to be assigned, or an object coercible by
\code{\link{getUniques}}.}

\item{refFasta}{(Required). A character string naming the path to the reference fasta file, or an 
R connection. This reference fasta file should be formatted with the id line corresponding to the
taxonomy (or classification) of the reference sequence, with each level separated by a semicolon. Eg.

 >Kingom;Phylum;Class;Order;Family;Genus;
 ACGAATGTGAAGTAA......}

\item{minBoot}{(Optional). Default 50. The minimum bootstrap confidence for assigning a taxonomic level.}

\item{verbose}{(Optional). \code{logical(1)} indicating verbose text output. Default FALSE.}
}
\value{
A character matrix of assigned taxonomies exceeding the minBoot level of
  bootstrapping confidence. Rows correspond to the provided sequences, columns to the
  taxonomic levels. NA indicates that the sequence was not consistently classified at
  that level at the minBoot threshhold.
}
\description{
assignTaxonomy implements the RDP classifier algorithm in Wang 2007 with kmer size 8 and
 100 bootstrap replicates.
}
\examples{
\dontrun{
 taxa <- assignTaxonomy(dadaF, "gg_13_8_97_dada.fa")
}

}

