% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_get_overlaps}
\alias{C_get_overlaps}
\title{Calculates the size of perfect overlap on the left and right side
of the child sequence (s2) to the aligned parent.}
\usage{
C_get_overlaps(s1, s2, max_shift)
}
\description{
Calculates the size of perfect overlap on the left and right side
of the child sequence (s2) to the aligned parent.
}

