% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_nwalign}
\alias{C_nwalign}
\title{Exposes ends-free Needleman-Wunsh alignment to R.}
\usage{
C_nwalign(s1, s2, score, gap_p, band)
}
\description{
Exposes ends-free Needleman-Wunsh alignment to R.
}

