% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_pair_consensus}
\alias{C_pair_consensus}
\title{Calculates the consensus of two sequences (first sequence wins mismatches).}
\usage{
C_pair_consensus(s1, s2)
}
\description{
Calculates the consensus of two sequences (first sequence wins mismatches).
}

