% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibrate_kmers}
\alias{calibrate_kmers}
\title{Generate the kmer-distance and the alignment distance from the
  given set of sequences.}
\usage{
calibrate_kmers(seqs, score, gap, max_aligns)
}
\arguments{
\item{seqs}{(Required). Character.
 A vector containing all unique sequences in the data set.
 Only A/C/G/T/N/- allowed. Ungapped sequences recommended.}

\item{score}{(Required). Numeric matrix (4x4).
The score matrix used during the alignment.}

\item{gap}{(Required). A \code{numeric(1)} giving the gap penalty for alignment.}

\item{max_aligns}{(Required). A \code{numeric(1)} giving the (maximum) number of
pairwise alignments to do.}
}
\value{
DataFrame.
}
\description{
Generate the kmer-distance and the alignment distance from the
  given set of sequences.
}

