% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiSample.R
\name{collapseNoMismatch}
\alias{collapseNoMismatch}
\title{Collapse together sequences that are identical up to shifts and/or length.}
\usage{
collapseNoMismatch(seqtab, minOverlap = 20, verbose = FALSE)
}
\arguments{
\item{seqtab}{(Required). A samples by sequence matrix, the return of \code{\link{makeSequenceTable}}.}

\item{minOverlap}{(Optional). \code{numeric(1)}. Default is 20.
The minimum amount of overlap between sequences required to collapse them together.}
}
\value{
Integer \code{matrix}.
A row for each sample, and a column for each collapsed sequence across all the samples.
Note that the columns are named by the sequence which can make display a little unwieldy.
Columns are in the same order (modulo the removed columns) as in the input matrix.
}
\description{
This function takes as input a sequence table and returns a sequence table in which
any sequences that are identical up to shifts or length variation, i.e. that have
no mismatches or internal indels when aligned, are collapsed together. The most abundant
sequences is chosen as the representative of the collapsed sequences.
}
\details{
This function can be thought of as implementing greedy 100-percent OTU clustering, where end-gapping
is ignored. The current implementation relies on full alignments and is therefore much slower
than necessary. A better implementation would be good.
}
\examples{
derep1 <- derepFastq(system.file("extdata", "sam1F.fastq.gz", package="dada2"))
derep2 <- derepFastq(system.file("extdata", "sam2F.fastq.gz", package="dada2"))
dada1 <- dada(derep1, tperr1)
dada2 <- dada(derep2, tperr1)
seqtab <- makeSequenceTable(list(sample1=dada1, sample2=dada2))
collapseNoMismatch(seqtab)
}
\seealso{
\code{\link{makeSequenceTable}}
}

