% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{dada-class}
\alias{dada-class}
\title{The object class returned by \code{\link{dada}}}
\description{
A multi-item List with the following named values...
\itemize{
 \item{$denoised: }{Integer vector, named by sequence valued by abundance, of the denoised sequences.}
 \item{$clustering: }{An informative data.frame containing information on each cluster.}
 \item{$sequence: }{A character vector of each denoised sequence. Identical to names($denoised).}
 \item{$quality: }{The average quality scores for each cluster (row) by position (col).}
 \item{$map: }{Integer vector that maps the unique (index of derep$unique) to the denoised sequence (index of dada$denoised).}
 \item{$birth_subs: }{A data.frame containing the substitutions at the birth of each new cluster.}
 \item{$trans: }{The matrix of transitions by type (row), eg. A2A, A2C..., and quality score (col)
         observed in the final output of the dada algorithm.}
 \item{$err_in: }{The err matrix used for this invocation of dada.}
 \item{$err_out: }{The err matrix estimated from the output of dada. NULL if err_function not provided.}
 \item{$opts: }{A list of the dada_opts used for this invocation of dada.}
 \item{$call: }{The function call used for this invocation of dada.}
}
}
\seealso{
\code{\link{dada}}
}

