% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dada.R
\name{dada}
\alias{dada}
\title{The dada function takes as input the unique sequences in an amplicon sequencing sample paired
 with their abundances, and returns the sample genotypes paired with their abundances
 inferred by the Divisive Amplicon Denoising Algorithm (Rosen, Callahan, Fisher, Holmes 2012).}
\usage{
dada(uniques, err = matrix(c(0.991, 0.003, 0.003, 0.003, 0.003, 0.991, 0.003,
  0.003, 0.003, 0.003, 0.991, 0.003, 0.003, 0.003, 0.003, 0.991), nrow = 4,
  byrow = T), self_consist = FALSE)
}
\arguments{
\item{uniques}{(Required). Named integer vector or list of named integer vectors.
 Each uniques vector is an integer vector of abundances, named by the unique DNA
   sequence to which that abundance corresponds.
 Sequences are only allowed to contain A/C/G/T characters.
 A list of such vectors can be provided, in which case the error model will be
   shared across these samples when denoising.}

\item{err}{(Optional). 4x4 numeric matrix.
 The matrix of estimated error rates from one nucleotide to another.
 err[i,j] = Prob(j in sequence | i in sample genotype).
 A=1, C=2, G=3, T=4 for indexing. Rows required to sum to 1.}

\item{self_consist}{(Optional). \code{logical(1)}
 When true, DADA will re-estimate error rates after inferring sample genotypes, and then repeat
 the algorithm using the newly estimated error rates. This continues until convergence.}
}
\value{
List.
 $genotypes: named integer vector of the denoised sample genotypes.
 $trans: 4x4 integer matrix of the inferred substutions ("errors") between nts.
 $opts: A list of the dada_opts used for this function call.
}
\description{
The dada function takes as input the unique sequences in an amplicon sequencing sample paired
 with their abundances, and returns the sample genotypes paired with their abundances
 inferred by the Divisive Amplicon Denoising Algorithm (Rosen, Callahan, Fisher, Holmes 2012).
}

