% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired.R
\name{dada_to_seq_table}
\alias{dada_to_seq_table}
\title{Map denoised sequence to each read.}
\usage{
dada_to_seq_table(dadaRes, derep, sr, idRegExpr = c("\\\\s.+$", ""),
  includeCol = character(0))
}
\arguments{
\item{dadaRes}{(Required). A \code{\link{dada-class}} object.
The output of \code{\link{dada}} function.}

\item{derep}{(Required).
A \code{\link{derep-class}} object.
An object returned by \code{\link{derepFastq}()}.
Should be the same object that served as input
to the \code{\link{dada}} function.}

\item{sr}{(Required). The trimmed and filtered reads 
that you used as input for \code{\link{derepFastq}},
prior to running \code{\link{dada}()} on \code{derep}.
More generally, this is an object that inherits from the 
\code{\link{ShortRead-class}}. 
In most cases this will be \code{\link{ShortReadQ-class}}.
Objects from this class are the result of \code{\link{readFastq}}.
Alternatively, this can be a character string
that provides the path to your forward reads fastq file.}

\item{idRegExpr}{(Optional). 
Exact same as for \code{\link{mergePairsByID}}.}

\item{includeCol}{(Optional). 
Exact same as for \code{\link{mergePairsByID}}.}
}
\value{
A \code{\link{data.table}}
in which each entry (row) contains
a unique read ID and the denoised sequence
to which it corresponds, as inferred by \code{\link{dada}}.
}
\description{
Takes the
\code{\link{dada}} result,
\code{\link{derepFastq}} result, and
\code{\link{ShortReadQ-class}} object
(an R representation of the original input sequence data),
and creates a \code{\link{data.table}}
in which each entry (row) contains
a unique read ID and the denoised sequence
to which it corresponds, as inferred by \code{\link{dada}}.
}
\examples{
exFileF = system.file("extdata", "sam1F.fastq.gz", package="dada2")
show(exFileF)
srF = ShortRead::readFastq(exFileF)
derepF = derepFastq(exFileF)
dadaF <- dada(derepF, err=tperr1,
              errorEstimationFunction=loessErrfun, selfConsist=TRUE)
dada2:::dada_to_seq_table(dadaF, derepF, srF)

}
\keyword{internal}

