% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dada_uniques}
\alias{dada_uniques}
\title{Run DADA on the provided unique sequences/abundance pairs.}
\usage{
dada_uniques(seqs, abundances, err, quals, score, gap, use_kmers, kdist_cutoff,
  band_size, omegaA, use_singletons, omegaS, maxClust, minFold, minHamming,
  useQuals, qMin, qMax)
}
\value{
List.
}
\description{
Run DADA on the provided unique sequences/abundance pairs.
}

