% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allPackage.R
\docType{package}
\name{dadac-package}
\alias{dadac-package}
\title{DADA package}
\description{
Some prose about the virtues of DADA here.
}
\author{
Benjamin Callahan \email{benjamin.j.callahan@gmail.com}

Paul J. McMurdie II \email{mcmurdie@stanford.edu}

Susan Holmes \email{susan@stat.stanford.edu}
}
\keyword{package}

