% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{derep-class}
\alias{derep-class}
\title{A class representing dereplicated sequences}
\description{
A \code{\link{list}} with the following three members.
\itemize{
 \item{$uniques: Named integer vector. Named by the unique sequence, valued by abundance.}
 \item{$quals: Numeric matrix of average quality scores by position for each unique. Uniques are rows, positions are cols.}
 \item{$map: Integer vector of length the number of reads, and value the index (in $uniques) of the unique to which that read was assigned.}
}
This can be created from a FastQ sequence file using
\code{\link{derepFastq}}
}
\seealso{
\code{\link{derepFastq}}
}

