% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{derepFasta}
\alias{derepFasta}
\title{derepFasta creates a derep-class object from a fasta file, by
creating a corresponding fastq file with a uniform quality score
and calling derepFastq.}
\usage{
derepFasta(fls, ...)
}
\arguments{
\item{fls}{(Required). \code{character}.
The file path(s) to the fasta or gzipped fasta file(s).}

\item{...}{(Optional).
Additional arguments passed on to \code{\link{derepFastq}}}
}
\description{
derepFasta creates a derep-class object from a fasta file, by
creating a corresponding fastq file with a uniform quality score
and calling derepFastq.
}
\keyword{internal}

