% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{dereplicateMultiSampleFastqReads}
\alias{dereplicateMultiSampleFastqReads}
\title{Read and Dereplicate a Fastq file containing multiple samples.}
\usage{
dereplicateMultiSampleFastqReads(fl, samsubseq, n = 1e+06, verbose = FALSE)
}
\arguments{
\item{fl}{(Required). Character.
 The file path to the fastq or fastq.gz file.}

\item{samsubseq}{(Required). A \code{numeric(2)} specifying the substring
 of the id field that will be used to split the fastq reads into samples.}

\item{n}{(Optional). A \code{numeric(1)} indicating
 the maximum number of records (reads) to parse and dereplicate
 at any one time. This controls the peak memory requirement.
 Defaults is \code{1e6}, one-million reads.
 See \code{\link{FastqStreamer}} for details on this parameter,
 which is passed on.}

\item{verbose}{(Optional). A \code{logical(1)} indicating
 whether to throw any standard R \code{\link{message}}s
 on the intermittent and final status of the dereplication.
 Default is \code{FALSE}, no messages.}
}
\value{
Named integer vector. Named by sequence, valued by number of occurence.
}
\description{
This is a custom interface to \code{\link{FastqStreamer}}
for dereplicating amplicon sequences from a multi-sample fastq or fastq.gz file.
This function relies heavily on the \code{\link[ShortRead]{tables}} method.
}
\seealso{
\code{\link{replicateReads}}, \code{\link{removeReadsWithNs}},
\code{\link{findBarcodes}}, \code{\link{splitByBarcode}}
}

