% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dereplicate.R
\name{dereplicateReads}
\alias{dereplicateReads}
\title{Removes duplicate sequences from DNAStringSet object.}
\usage{
dereplicateReads(dnaSet)
}
\arguments{
\item{dnaSet}{DNAStringSet object to dereplicate.}
}
\value{
DNAStringSet object with names describing frequency of repeat.
}
\description{
Borrowed from hiReadsProcessor package in BioC.
Given a DNAStringSet object, the function dereplicates reads and
adds counts=X to the definition line to indicate replication.
}
\examples{
dnaSet <- c("CCTGAATCCTGGCAATGTCATCATC", "ATCCTGGCAATGTCATCATCAATGG",
"ATCAGTTGTCAACGGCTAATACGCG", "ATCAATGGCGATTGCCGCGTCTGCA",
"CCGCGTCTGCAATGTGAGGGCCTAA", "GAAGGATGCCAGTTGAAGTTCACAC",
"CCTGAATCCTGGCAATGTCATCATC", "ATCCTGGCAATGTCATCATCAATGG",
"ATCAGTTGTCAACGGCTAATACGCG", "ATCAATGGCGATTGCCGCGTCTGCA",
"CCGCGTCTGCAATGTGAGGGCCTAA", "GAAGGATGCCAGTTGAAGTTCACAC")
dereplicateReads(dnaSet)
}
\seealso{
\code{\link{replicateReads}}, \code{\link{removeReadsWithNs}},
\code{\link{findBarcodes}}, \code{\link{splitByBarcode}}
}

