% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_band}
\alias{evaluate_band}
\title{Quantify the number of alignments altered by banding at the given BAND_SIZE.}
\usage{
evaluate_band(seqs, score, gap, band_size, max_aligns)
}
\arguments{
\item{seqs}{(Required). Character.
 A vector containing all unique sequences in the data set.
 Only A/C/G/T allowed.}

\item{score}{(Required). Numeric matrix (4x4).
The score matrix used during the alignment.}

\item{gap}{(Required). A \code{numeric(1)} giving the gap penalty for alignment.}

\item{band_size}{(Required). A \code{numeric(1)} giving the band size to consider.}

\item{max_aligns}{(Required). A \code{numeric(1)} giving the (maximum) number of
pairwise alignments to do.}
}
\value{
DataFrame.
}
\description{
Quantify the number of alignments altered by banding at the given BAND_SIZE.
}

