% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_kmers}
\alias{evaluate_kmers}
\title{Generate the kmer-distance and the alignment distance from the
  given set of sequences.}
\usage{
evaluate_kmers(seqs, kmer_size, score, gap, band, max_aligns)
}
\arguments{
\item{seqs}{(Required). Character.
 A vector containing all unique sequences in the data set.
 Only A/C/G/T allowed.}

\item{score}{(Required). Numeric matrix (4x4).
The score matrix used during the alignment. Coerced to integer.}

\item{gap}{(Required). A \code{numeric(1)} giving the gap penalty for alignment. Coerced to integer.}

\item{max_aligns}{(Required). A \code{numeric(1)} giving the (maximum) number of
pairwise alignments to do.}
}
\value{
DataFrame.
}
\description{
Generate the kmer-distance and the alignment distance from the
  given set of sequences.
}

