% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter.R
\name{fastqFilter}
\alias{fastqFilter}
\title{fastqFilter implements most of fastq_filter from usearch...}
\usage{
fastqFilter(fn, fout, maxN = 0, truncQ = "#", truncLen = 0,
  trimLeft = 0, minQ = 0, maxEE = Inf, n = 1e+06, compress = TRUE,
  verbose = FALSE)
}
\description{
fastqFilter implements most of fastq_filter from usearch...
}

