% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter.R
\name{filterFastq}
\alias{filterFastq}
\title{filterFastq implements most of fastq_filter from usearch...}
\usage{
filterFastq(fn, fout, maxN = 0, truncQ = "#", truncLen = 0,
  trimLeft = 0, minQ = 0, n = 1e+06, verbose = FALSE)
}
\description{
filterFastq implements most of fastq_filter from usearch...
}

