% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dadaOptions.R
\name{get_dada_opt}
\alias{get_dada_opt}
\title{Get DADA options}
\usage{
get_dada_opt(option = NULL)
}
\arguments{
\item{option}{(Optional). Character.
 The DADA options to get.}
}
\value{
Named list of option/value pairs.
 Returns NULL if an invalid option is requested.
}
\description{
Get DADA options
}

