% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errorModels.R
\name{getBadBases}
\alias{getBadBases}
\title{Identify bad base positions.}
\usage{
getBadBases(clust, subpos, omegaB = 1e-20, minOccurence = 4,
  verbose = FALSE)
}
\arguments{
\item{clust}{(Required). The $clustering data frame from the dada() output.
  May be subsetted from the original prior to using this function.}

\item{subpos}{(Required). The $subpos data frame from the dada() output.}

\item{omegaB}{(Optional). A \code{numeric(1)}. Default is 1e-10.
 The p-value threshhold below which a base is assigned as "bad".
 The p-value is calculated by the number of repeated occurences of a particular
   base position individually driving the formation of a new cluster. Bad bases
   drive many new "1-away" clustes.
 The null hypothesis being tested is that real differences are distributed
   uniformly along the sequence. This is not true, biological differences are
   non-uniform, so this pvalue threshhold should be set conservatively.}

\item{minOccurence}{(Optional). A \code{numeric(1)}. Default is 4.
 The minimum times a single base position must drive the formation of a new cluster
   before it can be considered a "bad base".}

\item{verbose}{(Optional). \code{logical(1)} indicating verbose text output. Defaults FALSE.}
}
\value{
Integer vector of the bad base positions.
}
\description{
Illumina sequencing sometimes produces "bad bases", positions at which
error rates are significantly higher than expected by the assigned quality
score. This function identifies those bad bases.
}
\seealso{
\code{\link{isBadBaseFP}}
}

