% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{getDerep}
\alias{getDerep}
\title{Get derep-class objects from the input object.}
\usage{
getDerep(object, ...)
}
\arguments{
\item{object}{(Required). The object from which to extract the \code{\link{derep-class}} object(s).}

\item{...}{(Optional). Arguments passed to \code{\link{derepFastq}} if it is called.}
}
\value{
A \code{\link{derep-class}} or a list of \code{\link{derep-class}} objects.
}
\description{
This function extracts a \code{\link{derep-class}} or list of \code{\link{derep-class}}
 from either an input \code{\link{derep-class}} or list of \code{\link{derep-class}} objects,
 or an input character vector of filenames to be dereplicated by \code{\link{derepFastq}}.
}
\examples{
fn <- system.file("extdata", "sam1F.fastq.gz", package="dada2")
derep1 = derepFastq(fn)
identical(getDerep(derep1), getDerep(fn))

}
