% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{getFasta}
\alias{getFasta}
\title{Read a FASTA file into a named uppercase character vector.}
\usage{
getFasta(fl)
}
\arguments{
\item{fl}{(Required). The path to the fasta file.}
}
\description{
A wrapper for readFasta in the ShortRead package.
}
