% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chimeras.R
\name{getOverlaps}
\alias{getOverlaps}
\title{Internal function that finds the best overlap between two sequences.
Uses NW alignment with ends-free gapping.}
\usage{
getOverlaps(parent, sq, allowOneOff = FALSE, maxShift = 16)
}
\description{
Internal function that finds the best overlap between two sequences.
Uses NW alignment with ends-free gapping.
}

