% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{getUniques}
\alias{getUniques}
\title{Get the vector of unique sequence.}
\usage{
getUniques(object)
}
\arguments{
\item{object}{(Required). The object from which to extract the uniques vector.}
}
\value{
\code{integer}.
 An integer vector named by unique sequence and valued by abundance.
}
\description{
This function extracts the uniques vector from several different data objects, including
 \code{\link{dada-class}}, \code{\link{derep-class}} and \code{data.frame} objects that have both
 $sequence and $abundance columns.
 The return value is of the "uniques vector": An integer vector that is named by the
 unique sequence and valued by abundance. If input is already a "uniques vector", that same vector
 will be returned.
}
\examples{
\dontrun{
 getUniques(dada.sample1)
 getUniques(merged.df)
 getUniques()
}
}

