% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errorModels.R
\name{inflateErr}
\alias{inflateErr}
\title{Function that inflates an error rate matrix by a specified factor. Accounts for saturation.
  new_err_rate <- err_rate * inflate / (1 + (inflate-1) * err_rate)}
\usage{
inflateErr(err, inflation = 1, inflateNonSubs = FALSE)
}
\description{
Function that inflates an error rate matrix by a specified factor. Accounts for saturation.
  new_err_rate <- err_rate * inflate / (1 + (inflate-1) * err_rate)
}

