% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errorModels.R
\name{isBadBaseFP}
\alias{isBadBaseFP}
\title{Identify False Positive inferred sequences due to bad bases.}
\usage{
isBadBaseFP(clust, subpos, minFraction = 0.51, omegaB = 1e-10,
  minOccurence = 4, verbose = FALSE)
}
\arguments{
\item{clust}{(Required). The $clustering data frame from the dada() output.
  May be subsetted from the original prior to using this function.}

\item{subpos}{(Required). The $subpos data frame from the dada() output.}

\item{minFraction}{(Optional). A \code{numeric(1)}. Default is 0.51.
 The minimum fraction of bad bases among the base positions used to infer the
 sequence required to call the inferred sequence a false positive.}

\item{omegaB}{(Optional). A \code{numeric(1)}. Default is 1e-10.
 The p-value threshhold below which a base is assigned as "bad".
 The p-value is calculated by the number of repeated occurences of a particular
   base position individually driving the formation of a new cluster. Bad bases
   drive many new "1-away" clustes.
 The null hypothesis being tested is that real differences are distributed
   uniformly along the sequence. This is not true, biological differences are
   non-uniform, so this pvalue threshhold should be set conservatively.}

\item{minOccurence}{(Optional). A \code{numeric(1)}. Default is 4.
 The minimum times a single base position must drive the formation of a new cluster
   before it can be considered a "bad base".}

\item{verbose}{(Optional). \code{logical(1)} indicating verbose text output. Default FALSE.}
}
\value{
Logical vector of length the number of inferred sequences.
 TRUE if inferred sequence a false positive.
 FALSE otherwise.
}
\description{
Illumina sequencing sometimes produces "bad bases", positions at which
error rates are significantly higher than expected by the assigned quality
score. This function identifies the inferred sequences that are likely to
have been driven by those bad bases.
}
\seealso{
\code{\link{getBadBases}}
}

