% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chimeras.R
\name{isBimera}
\alias{isBimera}
\title{Determine if input sequence is an exact bimera of putative parent sequences.}
\usage{
isBimera(sq, parents, allowOneOff = TRUE, minOneOffParentDistance = 4,
  maxShift = 16)
}
\arguments{
\item{sq}{(Required). A \code{character(1)}.
 The sequence being evaluated as a possible bimera.}

\item{parents}{(Required). Character vector.
 A vector of possible "parent" sequence that could form the left and right
 sides of the bimera.}
}
\value{
\code{logical(1)}.
 TRUE if sq is an exact bimera of two of the parents. Otherwise FALSE.
}
\description{
This function attempts to find an exact bimera of the parent sequences that
matches the input sequence. A bimera is a two-parent chimera, in which the
left side is made up of one parent sequence, and the right-side made up of
a second parent sequence. If an exact bimera is found (no mismatches are
allowed, but strict shifts are allowed) TRUE is returned, otherwise FALSE.
}
\seealso{
\code{\link{isBimeraDenovo}, \link{isBimeraOneOff}}
}

