% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chimeras.R
\name{isBimeraDenovo}
\alias{isBimeraDenovo}
\title{Identify bimeras de-novo from collections of unique sequences.}
\usage{
isBimeraDenovo(unqs, minFoldParentOverAbundance = 2, minParentAbundance = 8,
  allowOneOff = FALSE, minOneOffParentDistance = 4, maxShift = 16,
  verbose = FALSE)
}
\arguments{
\item{unqs}{(Required). A named integer vector or data.frame with "sequence" and "abundance" cols.
  This contains the sequences (and corresponding abundances) to be checked for bimeras.}

\item{minFoldParentOverAbundance}{(Optional). A \code{numeric(1)}. Default is 10.
  Only sequences at least this much-fold more abundant than a sequence can be its "parents".}

\item{minParentAbundance}{(Optional). A \code{numeric(1)}. Default is 100.
  Only sequences at least this abundant can be "parents".}

\item{allowOneOff}{(Optional). A \code{logical(1)}. Default is FALSE.
  If TRUE, sequences that have one mismatch or indel to an exact bimera are also
  flagged.}

\item{minOneOffParentDistance}{(Optional). A \code{numeric(1)}. Default is 4.
  Only sequences with at least this many mismatches to the potential bimeric sequence
  considered as possible "parents" when flagging one-off bimeras. Note that there is
  no such screen when considering exact bimeras.}

\item{maxShift}{(Optional). A \code{numeric(1)}. Default is 16.
  Maximum shift allowed when aligning sequences to potential "parents".}
}
\description{
This function is a wrapper around isBimera for collections of DADA denoised
sequences. Each sequence is evaluated against a set of "parents" drawn from the
sequence collection that are sufficiently more abundant than the sequence
being evaluated.
}
\seealso{
\code{\link{isBimera}}
}

