% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chimeras.R
\name{isShiftedPair}
\alias{isShiftedPair}
\title{Internal function that determines if two sequences are identical up to a shift
Uses NW alignment with ends-free gapping}
\usage{
isShiftedPair(sq1, sq2, minOverlap = 20)
}
\description{
Internal function that determines if two sequences are identical up to a shift
Uses NW alignment with ends-free gapping
}

