% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/errorModels.R
\name{makeErrfun}
\alias{makeErrfun}
\title{Make function that accepts the output transitions matrix from DADA ($subqual)
 and returns inferred error rates for each transition type (t_ij) and quality}
\usage{
makeErrfun(err_model, init_params, inflation = 1)
}
\description{
Make function that accepts the output transitions matrix from DADA ($subqual)
 and returns inferred error rates for each transition type (t_ij) and quality
}

