% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errorModels.R
\name{makeLoglsqObj}
\alias{makeLoglsqObj}
\title{Creates an objective function suitable for optim().
  Sums the square errors after taking the log transform.}
\usage{
makeLoglsqObj(err_model, qave, obs)
}
\description{
Creates an objective function suitable for optim().
  Sums the square errors after taking the log transform.
}

