% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiSample.R
\name{makeSequenceTable}
\alias{makeSequenceTable}
\title{Construct a sample-by-sequence observation matrix.}
\usage{
makeSequenceTable(samples)
}
\arguments{
\item{samples}{(Required). A \code{list}.
 The samples to include in the sequence table. Names are propagated to the rownames
 of the return matrix. Samples can be provided in any format that can
 be processed by \code{\link{getUniques}}.}
}
\value{
Integer \code{matrix}.
A row for each sample, and a column for each unique sequence across all the samples.
Note that the columns are named by the sequence which can make display a little unwieldy.
}
\description{
This function contructs a sequence table (analogous to an OTU table) from
the provided list of samples.
}
\examples{
derep1 <- derepFastq(system.file("extdata", "sam1F.fastq.gz", package="dada2"))
derep2 <- derepFastq(system.file("extdata", "sam2F.fastq.gz", package="dada2"))
dada1 <- dada(derep1, tperr1)
dada2 <- dada(derep2, tperr1)
makeSequenceTable(list(sample1=dada1, sample2=dada2))
}
\seealso{
\code{\link{dada}}, \code{\link{getUniques}}
}

