% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{makeSpeciesFasta_Silva}
\alias{makeSpeciesFasta_Silva}
\title{This function creates the dada2 assignSpecies fasta file for Silva
from the SILVA_[VERSION]_SSURef_tax_silva.fasta file (NOT the NR99 file).}
\usage{
makeSpeciesFasta_Silva(fin, fout, compress = TRUE)
}
\description{
## Silva release v138.2
dada2:::makeSpeciesFasta_Silva("~/tax/silva/v138_2/SILVA_138.2_SSURef_tax_silva.fasta.gz", 
    "~/Desktop/silva_v138.2_assignSpecies.fa.gz")
}
\details{
Output: 352047 sequences with genus/species binomial annotation output.
}
\keyword{internal}
