% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{makeTaxonomyFasta_RDP}
\alias{makeTaxonomyFasta_RDP}
\title{This function creates the dada2 assignTaxonomy training fasta for the RDP
from the trainset16_022016.fa file.}
\usage{
makeTaxonomyFasta_RDP(fin, fdb, fout, compress = TRUE)
}
\description{
This function creates the dada2 assignTaxonomy training fasta for the RDP
from the trainset16_022016.fa file.
}
\keyword{internal}
