% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{makeTaxonomyFasta_Silva}
\alias{makeTaxonomyFasta_Silva}
\title{This function creates the dada2 assignTaxonomy training fasta for the Silva .align file
generated by the Mothur project.}
\usage{
makeTaxonomyFasta_Silva(fin, ftax, fout, compress = TRUE)
}
\description{
## Silva release v128
path <- "~/Desktop/Silva/Silva.nr_v128"
dada2:::makeTaxonomyFasta_Silva(file.path(path, "silva.nr_v128.align"), 
    file.path(path, "silva.nr_v128.tax"), 
    "~/tax/silva_nr_v128_train_set.fa.gz")
}
\details{
## Silva release v132
path <- "~/Desktop/Silva.nr_v128"
dada2:::makeTaxonomyFasta_Silva(file.path(path, "silva.nr_v132.align"), 
    file.path(path, "silva.nr_v132.tax"), 
    "~/tax/silva_nr_v132_train_set.fa.gz")
}
\keyword{internal}
