% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{names<-,derep,ANY-method}
\alias{names<-,derep,ANY-method}
\title{Deactivate renaming of derep-class objects.}
\usage{
\S4method{names}{derep,ANY}(x) <- value
}
\arguments{
\item{x}{an \R object.}

\item{value}{a character vector of up to the same length as \code{x}, or
    \code{NULL}.}
}
\value{
NULL.
}
\description{
Deactivate renaming of derep-class objects.
}
