% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nwalign}
\alias{nwalign}
\title{Needlman-Wunsch alignment with ends-free gapping.}
\usage{
nwalign(s1, s2, score = getDadaOpt("SCORE_MATRIX"),
  gap = getDadaOpt("GAP_PENALTY"), band = getDadaOpt("BAND_SIZE"),
  endsfree = TRUE)
}
\arguments{
\item{s1}{(Required). \code{character(1)}. The first sequence to align. A/C/G/T only.}

\item{s2}{(Required). \code{character(1)}. The second sequence to align. A/C/G/T only.}

\item{score}{(Optional). A 4x4 numeric matrix.
The transition scores to use for the alignment. Default is getDadaOpt("SCORE_MATRIX").}

\item{gap}{(Optional). A \code{numeric(1)}. The gap penalty.
Default is getDadaOpt("GAP_PENALTY").}

\item{band}{(Optional). A \code{numeric(1)}. The band size.
This Needleman-Wunsch alignment is banded. This value specifies the radius of that band.
Default is getDadaOpt("BAND_SIZE"). Set band = -1 for unbanded alignment.}

\item{endsfree}{(Optional). \code{logical(1)}. Default TRUE.
Ends-free gapping (or not).}
}
\value{
\code{character(2)}. The aligned sequences.
}
\description{
This function performs a Needleman-Wunsch alignment with free gaps at the ends of the sequences.
}
\examples{
 sq1 <- "CTAATACATGCAAGTCGAGCGAGTCTGCCTTGAAGATCGGAGTGCTTGCACTCTGTGAAACAAGATA"
 sq2 <- "TTAACACATGCAAGTCGAACGGAAAGGCCAGTGCTTGCACTGGTACTCGAGTGGCGAACGGGTGAGT"
 nwalign(sq1, sq2)
 nwalign(sq1, sq2, band=-1)

}

