% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{nwhamming}
\alias{nwhamming}
\title{Hamming distance after Needlman-Wunsch alignment with ends-free gapping.}
\usage{
nwhamming(s1, s2, ...)
}
\arguments{
\item{s1}{(Required). \code{character(1)}. The first sequence to align. A/C/G/T only.}

\item{s2}{(Required). \code{character(1)}. The second sequence to align. A/C/G/T only.}

\item{...}{(Optional). Further arguments to pass on to \code{\link{nwalign}}.}
}
\value{
\code{integer(1)}. The total number of mismatches and gaps, excluding gaps at the beginning and end of the alignment.
}
\description{
This function performs a Needleman-Wunsch alignment with free gaps at the ends of the sequences, and then counts
the number of mismatches and indels in that alignment. Gaps at the beginning and end are ignored.
}
\examples{
sq1 <- "CTAATACATGCAAGTCGAGCGAGTCTGCCTTGAAGATCGGAGTGCTTGCACTCTGTGAAACAAGATACAGGCTAGCGGCGGACGGGTGAGTAACACGTGGGTAACCTGCCCAAGAGATCGGGATAACACCTGGAAACAGATGCTAATACCGGATAACAACAGATGATGCCTATCAACTGTTTAAAAGATGGTTCTGCTATCACTCTTGGATGGACCTGCG"
 sq2 <- "TTAACACATGCAAGTCGAACGGAAAGGCCAGTGCTTGCACTGGTACTCGAGTGGCGAACGGGTGAGTAACACGTGGGTGATCTGCCCTGTACTTCGGGATAAGCTTGGGAAACTGGGTCTAATACCGGATAGGACAACTTTTTGGATATTGTTGTGGAAAGCTTTTGCGGTATGGGATGAGCTCGCGGCCTATCAGCTTGTTGGTGGGGTAATGGCCTAC"
 nwhamming(sq1, sq2)
 nwhamming(sq1, sq2, band=-1)
}

