% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_substitutions}
\alias{plot_substitutions}
\title{Plot Substitution Pairs from DADA Result}
\usage{
plot_substitutions(dadaOut, facetByGrp = TRUE)
}
\arguments{
\item{dadaOut}{(Required). The object returned by \code{\link{dada}}.}

\item{facetByGrp}{(Optional). Logical(1).
 Whether to plot all substitution groups together in one panel
 or separately on a grid of panels with a linear model fit.}
}
\value{
A \code{\link{ggplot}2} object that will be rendered
 to default device if \code{\link{print}ed},
 or can be stored and further modified.
 See \code{\link{ggsave}} for additional options.
}
\description{
This is similar to original DADA article, Figure 6.
}
\examples{
# Examples here.
testFile = system.file("extdata", "test-nonunique.fastq.gz", package="dadac")
test1 = derepFastq(testFile, verbose = TRUE)
test1$quals[test1$quals > 40] <- 40
res1 <- dada(uniques = test1$uniques, quals = test1$quals,
             err = dadac:::inflateErr(tperr1, 2),
             OMEGA_A = 1e-40,
             USE_QUALS = TRUE,
             err_function = makeErrfun(threepiece, dadac:::tp_init_parms),
             self_consist = TRUE)
plot_substitutions(res1)
}

