% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plotComplementarySubstitutions}
\alias{plotComplementarySubstitutions}
\title{Plot Substitution Pairs from DADA Result}
\usage{
plotComplementarySubstitutions(dadaOut, facetByGrp = TRUE)
}
\arguments{
\item{dadaOut}{(Required). A \code{\link{dada-class}} object.}

\item{facetByGrp}{(Optional). Default TRUE.
Whether to plot all substitution groups together in one panel
or separately on a grid of panels with a linear model fit.}
}
\value{
A \code{\link{ggplot}2} object.
 Will be rendered to default device if \code{\link{print}ed},
 or can be stored and further modified.
 See \code{\link{ggsave}} for additional options.
}
\description{
This is similar to original DADA article, Figure 6.
}
\examples{
derep1 = derepFastq(system.file("extdata", "sam1F.fastq.gz", package="dada2"), verbose = TRUE)
dada1 <- dada(derep1, err = inflateErr(tperr1, 2), selfConsist = TRUE) 
plotComplementarySubstitutions(dada1)

}
