% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plotQualityProfile}
\alias{plotQualityProfile}
\title{Plot quality profile of a fastq file.}
\usage{
plotQualityProfile(fl)
}
\arguments{
\item{fl}{(Required). \code{character(1)}.
The file path to the fastq or fastq.gz file.}
}
\value{
A \link{ggplot} object.
}
\description{
This function plots a visual summary of the distribution of quality scores
along the sequences within a fastq file. This function returns a \link{ggplot}
graphics object, which can be further modified in ggplot2 fashion.
}
\examples{
plotQualityProfile(system.file("extdata", "sam1F.fastq.gz", package="dada2")) + ggtitle("Test")

}

