% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{qtables}
\alias{qtables}
\title{Internal function to replicate tables functionality while also returning ave quals}
\usage{
qtables(x)
}
\description{
Internal function to replicate tables functionality while also returning ave quals
}

