% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{qtables2}
\alias{qtables2}
\title{Internal function to replicate tables functionality while also returning average quals and a map
from reads to uniques}
\usage{
qtables2(x, QtoP = FALSE)
}
\description{
Internal function to replicate tables functionality while also returning average quals and a map
from reads to uniques
}

