% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{showErrors}
\alias{showErrors}
\title{Plot error rates from after dada processing.}
\usage{
showErrors(dq, nti, ntj = "all", erri = TRUE, erro = TRUE, ...)
}
\arguments{
\item{dq}{Required. A dada return object.}

\item{nti}{Required. The base from which errors are displayed.}

\item{ntj}{Optional. Default "all". Choice of "A", "C", "G", "T".
 The error rate from nti->ntj will be plotted. If "all" is selected,
 the error rate from nti to all 4 other bases will be plotted.}

\item{erri}{Optional. Default TRUE.
 A \code{logical(1)} determining whether to plot the input error rates.}

\item{erro}{Optional. Default TRUE.
 A \code{logical(1)} determining whether to plot the output error rates.}
}
\value{
A \code{\link{ggplot2}} object that will be rendered
 to default device if \code{\link{print}ed},
 or can be stored and further modified.
}
\description{
This function plots the observed freuency of substitutions for each transition
(eg. A->C) as a function of the associated quality score. It also plots the error
raates passed in to this instantiation of the dada algorithm, and the output error
rates if they exist.
}

