% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/errorModels.R
\name{threepiece}
\alias{threepiece}
\title{An error model for the dependence of error rates on quality in Illumina data.}
\usage{
threepiece(parms, qave)
}
\description{
A piecewise linear model with three segments. The first and third segments are flat.
}

