% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{uniques-vector}
\alias{uniques-vector}
\title{The named integer vector format used to represent collections of unique DNA sequences.}
\description{
The uniques vector is an \code{integer} vector that is named by the unique sequence, and 
valued by the abundance of that sequence. This format is commonly used within the 
\code{\link{dada2-package}}, for function inputs and outputs. The \code{\link{getUniques}}
function coerces a variety of input objects into the uniques-vector format, including
\code{\link{dada-class}} and \code{\link{derep-class}} objects.
}
\seealso{
\code{\link{getUniques}}
}
