% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceIO.R
\docType{methods}
\name{writeFasta,character-method}
\alias{writeFasta,character-method}
\title{Writes a named character vector of DNA sequences to a fasta file.
Values are the sequences, and names are used for the id lines.}
\usage{
\S4method{writeFasta}{character}(object, file, mode = "w", width = 20000L,
  ...)
}
\arguments{
\item{object}{(Required). A named \code{character} vector.}

\item{file}{(Required). The output file.}

\item{mode}{(Optional). Default "w". Append with "a".}

\item{width}{(Optional). Default 20000L. Maximum line length before newline.}

\item{...}{(Optional). Additonal arguments passed to \code{\link[Biostrings]{writeXStringSet}}.}
}
\value{
NULL.
}
\description{
Writes a named character vector of DNA sequences to a fasta file.
Values are the sequences, and names are used for the id lines.
}
\seealso{
\code{\link[Biostrings]{writeXStringSet}}
}
