% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{c,derep-method}
\alias{c,derep-method}
\title{Change concatenation of derep-class objects to list construction.}
\usage{
\S4method{c}{derep}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{A derep-class object}

\item{...}{objects to be concatenated.  All \code{\link[base]{NULL}} entries
    are dropped before method dispatch unless at the very beginning of the
    argument list.}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}
}
\value{
list.
}
\description{
Change concatenation of derep-class objects to list construction.
}
