% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_nwalign}
\alias{C_nwalign}
\title{Exposes ends-free Needleman-Wunsh alignment to R.}
\usage{
C_nwalign(s1, s2, score, gap_p, band)
}
\arguments{
\item{s1}{A length-1 character verctor of sequence 1.}

\item{s2}{A length-1 character verctor of sequence 2.}

\item{score}{The 4x4 score matrix for nucleotide transitions.}

\item{gap_p}{The gap penalty.}

\item{band}{The band size (-1 turns off banding).}
}
\description{
Exposes ends-free Needleman-Wunsh alignment to R.
}

