% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dada_uniques}
\alias{dada_uniques}
\title{Run DADA on the provided unique sequences/abundance pairs.}
\usage{
dada_uniques(seqs, abundances)
}
\arguments{
\item{seqs}{(Required). Character.
 A vector containing all unique sequences in the data set.
 Only A/C/G/T/N/- allowed. Ungapped sequences recommended.}

\item{abundances}{(Required). Numeric.
 A vector of the number of reads of each unique seuqences.
 NAs not tolerated. Must be same length as the seqs vector.}
}
\value{
DataFrame object with sequence and abundance columns,
corresponding to the DADA denoised sample genotypes.
}
\description{
Run DADA on the provided unique sequences/abundance pairs.
}

