% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{derepFastq}
\alias{derepFastq}
\title{Read and Dereplicate a Fastq file.}
\usage{
derepFastq(fl, n = 1e+06, verbose = FALSE)
}
\arguments{
\item{fl}{(Required). Character.
 The file path to the fastq or fastq.gz file.
 Actually, any file format supported by \code{\link{FastqStreamer}}.}

\item{n}{(Optional). A \code{numeric(1)} indicating
 the maximum number of records (reads) to parse and dereplicate
 at any one time. This controls the peak memory requirement
 so that large fastq files are supported.
 Defaults is \code{1e6}, one-million reads.
 See \code{\link{FastqStreamer}} for details on this parameter,
 which is passed on.}

\item{verbose}{(Optional). A \code{logical(1)} indicating
 whether to throw any standard R \code{\link{message}}s
 on the intermittent and final status of the dereplication.
 Default is \code{FALSE}, no messages.}
}
\value{
List.
 $uniques: Named integer vector. Named by the unique sequence, valued by abundance.
 $quals: Numeric matrix of average quality scores by position for each unique. Uniques are rows, positions are cols.
 $map: Integer vector of the index of the unique (in $uniques) corresponding to each read.
}
\description{
This is a custom interface to \code{\link{FastqStreamer}}
for dereplicating amplicon sequences from a fastq or compressed fastq file,
while also controlling peak memory requirement to support large files.
This function relies heavily on the \code{\link[ShortRead]{tables}} method.
}
\examples{
# Test that chunk-size, `n`, does not affect the result.
testFile = system.file("extdata", "test-nonunique.fastq.gz", package="dadac")
test1 = dereplicateFastqReads(testFile, verbose = TRUE)
test2 = dereplicateFastqReads(testFile, 35, TRUE)
test3 = dereplicateFastqReads(testFile, 100, TRUE)
all.equal(test1, test2[names(test1)])
all.equal(test1, test3[names(test1)])
}
\seealso{
\code{\link{replicateReads}}
}

