% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{derepFastqWithQual}
\alias{derepFastqWithQual}
\title{Read and Dereplicate a Fastq file. Also returns average Q score for each unique.}
\usage{
derepFastqWithQual(fl, n = 1e+06, verbose = FALSE)
}
\description{
Read and Dereplicate a Fastq file. Also returns average Q score for each unique.
}

