% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_kmers}
\alias{evaluate_kmers}
\title{Generate the kmer-distance and the alignment distance from the
  given set of sequences.}
\usage{
evaluate_kmers(seqs, kmer_size, score, gap, band, max_aligns)
}
\arguments{
\item{seqs}{(Required). Character.
A vector containing all unique sequences in the data set.
Only A/C/G/T allowed.}

\item{kmer_size}{(Required). A \code{numeric(1)}. The size of the kmer to test (eg. 5-mer).}

\item{score}{(Required). Numeric matrix (4x4).
The score matrix used during the alignment. Coerced to integer.}

\item{gap}{(Required). A \code{numeric(1)} giving the gap penalty for alignment. Coerced to integer.}

\item{band}{(Required). A \code{numeric(1)} giving the band-size for the NW alignments.}

\item{max_aligns}{(Required). A \code{numeric(1)} giving the (maximum) number of
pairwise alignments to do.}
}
\value{
DataFrame.
}
\description{
Generate the kmer-distance and the alignment distance from the
  given set of sequences.
}
\examples{
derep1 = derepFastq(system.file("extdata", "sam1F.fastq.gz", package="dada2"))
kmerdf <- dada2:::evaluate_kmers(getSequences(derep1), 5, getDadaOpt("SCORE_MATRIX"),
                                 getDadaOpt("GAP_PENALTY"), 16, 1000)
plot(kmerdf$kmer, kmerdf$align)

}

