% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{importUniques}
\alias{importUniques}
\title{Load .uniques file}
\usage{
importUniques(fl, colClasses = c("integer", "character"), ...)
}
\arguments{
\item{fl}{(Required). \code{character(1)}.
 The file path to the .uniques file,
 a delimited text table file containing unique sequences and their abundances.}

\item{colClasses}{(Optional). \code{character}.
 The classes of the columns in the delimited text file.
 Defaults to \code{c("integer", "character")}.
 See \code{\link[utils]{read.table}} for more information.}

\item{...}{(Optional). Additional arguments passed on to \code{\link[utils]{read.table}}.}
}
\value{
Named integer vector. Named by sequence, valued by abundance.
}
\description{
This is a custom interface for read.table customized for loading .uniques format
 files. The .uniques format is a delimited text file (tab-delimited by default)
 with the abundance in column 1 and the ASCII sequence in column 2.
}

