% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{importUniques}
\alias{importUniques}
\title{Load .uniques file
Basically a wrapper for read.table customized for .uniques format}
\usage{
importUniques(fl, sep = "")
}
\arguments{
\item{fl}{(Required). Character.
 The file path to the .uniques file.}

\item{sep}{(Optional). The field separator character.}
}
\value{
Named integer vector. Named by sequence, valued by number of occurence.
}
\description{
Load .uniques file
Basically a wrapper for read.table customized for .uniques format
}

