% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeras.R
\name{isBimeraDenovo}
\alias{isBimeraDenovo}
\title{Identify bimeras de-novo from collections of unique sequences.}
\usage{
isBimeraDenovo(unqs, minFoldParentOverAbundance = 1, minParentAbundance = 8,
  allowOneOff = TRUE, minOneOffParentDistance = 4, maxShift = 16,
  verbose = FALSE)
}
\arguments{
\item{unqs}{(Required). A "uniques vector" or any object that can be coerced into one with \code{\link{getUniques}}.
This named integer vector is named by the sequences to be checked for bimeras and valued by their abundances.}

\item{minFoldParentOverAbundance}{(Optional). A \code{numeric(1)}. Default is 1.
Only sequences more than this-fold abundant than a sequence can be its "parents".
Default is intentionally permissive, as aggressively removing chimeras
 is the conservative choice for downstream analysis.}

\item{minParentAbundance}{(Optional). A \code{numeric(1)}. Default is 8.
Only sequences at least this abundant can be "parents".}

\item{allowOneOff}{(Optional). A \code{logical(1)}. Default is TRUE.
If TRUE, sequences that have one mismatch or indel to an exact bimera are also
flagged.}

\item{minOneOffParentDistance}{(Optional). A \code{numeric(1)}. Default is 4.
Only sequences with at least this many mismatches to the potential bimeric sequence
considered as possible "parents" when flagging one-off bimeras. Note that there is
no such screen when considering exact bimeras.}

\item{maxShift}{(Optional). A \code{numeric(1)}. Default is 16.
Maximum shift allowed when aligning sequences to potential "parents".}

\item{verbose}{(Optional). \code{logical(1)} indicating verbose text output. Default FALSE.}
}
\value{
\code{logical} of length the number of input unique sequences.
 TRUE if sequence is a bimera of more abundant "parent" sequences. Otherwise FALSE.
}
\description{
This function is a wrapper around isBimera for collections of DADA denoised
sequences. Each sequence is evaluated against a set of "parents" drawn from the
sequence collection that are sufficiently more abundant than the sequence
being evaluated.
}
\examples{
derep1 = derepFastq(system.file("extdata", "sam1F.fastq.gz", package="dada2"))
dada1 <- dada(derep1, err=tperr1, errorEstimationFunction=loessErrfun, selfConsist=TRUE)
isBimeraDenovo(dada1)
isBimeraDenovo(getUniques(dada1), minFoldParentOverAbundance = 2, allowOneOff=FALSE)

}
\seealso{
\code{\link{isBimera}}
}

