% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chimeras.R
\name{isShiftDenovo}
\alias{isShiftDenovo}
\title{Identify sequences that are identical to a more abundant sequence up to an
overall shift from a collection of unique sequences.}
\usage{
isShiftDenovo(unqs, minOverlap = 20, verbose = FALSE)
}
\arguments{
\item{unqs}{(Required). A named integer vector or data.frame with "sequence" and "abundance" cols.
  This contains the sequences (and corresponding abundances) to be checked for shifts.}

\item{minOverlap}{(Optional). A \code{numeric(1)}. Default is 20.
  Minimum overlap required to call something a shift.}
}
\value{
\code{logical} of length the number of input unique sequences.
 TRUE if sequence is an exact shift of a more abundant sequence. Otherwise FALSE.
}
\description{
This function is a wrapper around isShift for collections of DADA denoised
sequences. Each sequence is evaluated against a set of "parents" drawn from the
sequence collection that are more abundant than the sequence being evaluated.
}
\seealso{
\code{\link{isBimera}}
}

