% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{makeSpeciesFasta_RDP}
\alias{makeSpeciesFasta_RDP}
\title{This function creates the dada2 assignSpecies fasta file for the RDP
from the rdp_Bacteria_unaligned.fa file.}
\usage{
makeSpeciesFasta_RDP(fin, fout, compress = TRUE)
}
\description{
This function creates the dada2 assignSpecies fasta file for the RDP
from the rdp_Bacteria_unaligned.fa file.
}
\keyword{internal}

