% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{makeSpeciesFasta_RDP}
\alias{makeSpeciesFasta_RDP}
\title{This function creates the dada2 assignSpecies fasta file for the RDP
from the RDP's _Bacteria_unaligned.fa file.}
\usage{
makeSpeciesFasta_RDP(fin, fout, compress = TRUE)
}
\description{
THE WAY RDP RELEASES SPECIES LEVEL INFORMATION APPEARS TO HAVE CHANGED IN RELEASE 19
AS A RESULT, THIS OUTPUT IS NOT CURRENTLY BEING MAINTAINED
}
\details{
## RDP Trainset 18/Release 11.5
## The RDP documentation does not make clear whether the updates to the taxonomy from training set release 18 were
## propagated to the current Bacterial alignment.
dada2:::makeSpeciesFasta_RDP("~/Desktop/RDP/current_Bacteria_unaligned.fa", "~/tax/rdp_species_assignment_18.fa.gz")
dada2:::tax.check("~/tax/rdp_species_assignment_18.fa.gz", mode="species")

## RDP Trainset 16/Release 11.5
dada2:::makeSpeciesFasta_RDP("~/Desktop/RDP/current_Bacteria_unaligned.fa", "~/tax/rdp_species_assignment_16.fa.gz")
}
\keyword{internal}
