% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{makeSpeciesFasta_Silva}
\alias{makeSpeciesFasta_Silva}
\title{This function creates the dada2 assignSpecies fasta file for Silva
from the SILVA_123.1_SSURef_3_3_2016_fix.fasta file.}
\usage{
makeSpeciesFasta_Silva(fin, fout, compress = TRUE)
}
\description{
This function creates the dada2 assignSpecies fasta file for Silva
from the SILVA_123.1_SSURef_3_3_2016_fix.fasta file.
}
\keyword{internal}

