% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{makeTaxonomyFasta_SilvaNR}
\alias{makeTaxonomyFasta_SilvaNR}
\title{This function creates the dada2 assignTaxonomy training fasta for the official Silva NR99
release files. If `include.species`=TRUE, a 7th taxonomic level (species) will be added based on the
Genus species binomial in the Silva taxonomy string, if present and valid.}
\usage{
makeTaxonomyFasta_SilvaNR(
  fin,
  ftax,
  fout,
  include.species = FALSE,
  compress = TRUE
)
}
\description{
## Silva release v138
path <- "~/tax/Silva/v138"
dada2:::makeTaxonomyFasta_SilvaNR(file.path(path, "SILVA_138_SSURef_NR99_tax_silva.fasta.gz"), 
    file.path(path, "tax_slv_ssu_138.txt"), 
    "~/Desktop/silva_nr99_v138_train_set.fa.gz")
dada2:::makeTaxonomyFasta_SilvaNR(file.path(path, "SILVA_138_SSURef_NR99_tax_silva.fasta.gz"), 
    file.path(path, "tax_slv_ssu_138.txt"), 
    include.species=TRUE, "~/Desktop/silva_nr99_v138_wSpecies_train_set.fa.gz")
}
\keyword{internal}
