% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paired.R
\name{mergePairs}
\alias{mergePairs}
\title{Merge paired forward and reverse reads after DADA denoising.}
\usage{
mergePairs(dadaF, mapF, dadaR, mapR, keepMismatch = FALSE, minOverlap = 20,
  propagateCol = character(0), verbose = TRUE)
}
\arguments{
\item{dadaF}{(Required). Output of dada() function.
 The output list returned by the dada() function on the forward reads.}

\item{mapF}{(Required). An integer vector map from read index to unique index.
 The map returned by derepFastq() for the forward reads.}

\item{dadaR}{(Required). Output of dada() function.
 The output list returned by the dada() function on the reverse reads.}

\item{mapR}{(Required). An integer vector map from read index to unique index.
 The map returned by derepFastq() for the reverse reads.}

\item{minOverlap}{(Optional). A \code{numeric(1)} of the minimum overlap
 required for merging the forward and reverse reads. Default is 20.}

\item{propagateCol}{(Optional). Character vector. Default is empty.
 The mergePairs return data.frame will include copies of columns with names specified
 in the dada()$clustering data.frame.}

\item{verbose}{(Optional). \code{logical(1)} indicating verbose text output. Default FALSE.}

\item{keepMismatch(Optional).}{A \code{logical(1)}. Default is False.
 If true, the pairs that did not match are retained in the return data.frame.}
}
\value{
Dataframe. One row for each unique pairing of forward/reverse denoised sequences.
\itemize{
 \item{$abundance: }{Number of reads corresponding to this forward/reverse combination.}
 \item{$sequence: The merged sequence if match=TRUE. Otherwise an empty sequence, i.e. "".}
 \item{$forward: The index of the forward denoised genotypes.}
 \item{$reverse: The index of the reverse denoised genotype.}
 \item{$nmatch: Number of matching nts in the overlap region.}
 \item{$nmismatch: Number of mismatching nts in the overlap region.}
 \item{$nindel: Number of indels in the overlap region.}
 \item{$match: TRUE if a perfect match between the forward and reverse denoised sequences of at least MIN_OVERLAP.
         FALSE otherwise.}
 \item{$...: Additional columns specified in the propagateCol argument.}
}
}
\description{
This function takes the output of dada() and the map from read to unique index
returned by derepFastq()$map for both the forward and reverse data set. It attempts
to merge each pair of reads, rejecting any which do not perfectly overlap.
Note: This function assumes that the fastq files for the forward and reverse reads
were in the same order.
}
\seealso{
\code{\link{derepFastq}}, \code{\link{dada}}
}

