% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paired.R
\name{mergePairs}
\alias{mergePairs}
\title{Function to merge DADA-denoised paired end reads}
\usage{
mergePairs(dqF, mapF, dqR, mapR, minOverlap = 20)
}
\description{
Function to merge DADA-denoised paired end reads
}

