% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nwalign}
\alias{nwalign}
\title{Needlman-Wunsch alignment.}
\usage{
nwalign(s1, s2, score = getDadaOpt("SCORE_MATRIX"),
  gap = getDadaOpt("GAP_PENALTY"), homo_gap = NULL, band = -1,
  endsfree = TRUE)
}
\arguments{
\item{s1}{(Required). \code{character(1)}. The first sequence to align. A/C/G/T only.}

\item{s2}{(Required). \code{character(1)}. The second sequence to align. A/C/G/T only.}

\item{score}{(Optional). A 4x4 numeric matrix. Default is getDadaOpt("SCORE_MATRIX").
The match/mismatch used for the alignment.}

\item{gap}{(Optional). \code{numeric(1)}. Default is getDadaOpt("GAP_PENALTY").
The alignment gap penalty. Should be negative.}

\item{homo_gap}{(Optional). \code{numeric(1)}. Default NULL (no special homopolymer penalty).
The alignment gap penalty within homopolymer regions. Should be negative.}

\item{band}{(Optional). \code{numeric(1)}.  Default -1 (no banding).
This Needleman-Wunsch alignment can be banded. This value specifies the radius of that band.
Set \code{band = -1} to turn off banding.}

\item{endsfree}{(Optional). \code{logical(1)}. Default TRUE.
Allow free gapping at the ends of sequences.}
}
\value{
\code{character(2)}. The aligned sequences.
}
\description{
This function performs a Needleman-Wunsch alignment between two sequences.
}
\examples{
 sq1 <- "CTAATACATGCAAGTCGAGCGAGTCTGCCTTGAAGATCGGAGTGCTTGCACTCTGTGAAACAAGATA"
 sq2 <- "TTAACACATGCAAGTCGAACGGAAAGGCCAGTGCTTGCACTGGTACTCGAGTGGCGAACGGGTGAGT"
 nwalign(sq1, sq2)
 nwalign(sq1, sq2, band=16)

}

