% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plotComplementarySubstitutions}
\alias{plotComplementarySubstitutions}
\title{Plot Substitution Pairs from DADA Result}
\usage{
plotComplementarySubstitutions(dadaOut, facetByGrp = TRUE)
}
\arguments{
\item{dadaOut}{(Required). The object returned by \code{\link{dada}}.}

\item{facetByGrp}{(Optional). Logical(1).
 Whether to plot all substitution groups together in one panel
 or separately on a grid of panels with a linear model fit.}
}
\value{
A \code{\link{ggplot}2} object that will be rendered
 to default device if \code{\link{print}ed},
 or can be stored and further modified.
 See \code{\link{ggsave}} for additional options.
}
\description{
This is similar to original DADA article, Figure 6.
}
\examples{
derep1 = derepFastq(system.file("extdata", "sam1F.fastq.gz", package="dada2"), verbose = TRUE)
dada1 <- dada(derep1, err = inflateErr(tperr1, 2), errorEstimationFunction = loessErrfun, selfConsist = TRUE)
plotComplementarySubstitutions(dada1)
}

