% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plotErrors}
\alias{plotErrors}
\title{Plot error rates from after dada processing.}
\usage{
plotErrors(dq, nti = "all", ntj = "all", err_out = TRUE, err_in = FALSE,
  nominal_q = FALSE, ...)
}
\arguments{
\item{dq}{Required. A dada return object.}

\item{nti}{Optional. Default "all". Choice of "A", "C", "G", "T".}

\item{ntj}{Optional. Default "all". Choice of "A", "C", "G", "T".
 The error rate from nti->ntj will be plotted. If "all" is selected,
 the error rate from all 4 nti, or to all 4 ntj, will be plotted.}

\item{err_out}{Optional. Default TRUE.
 A \code{logical(1)} determining whether to plot the output error rates (solid).}

\item{err_in}{Optional. Default FALSE.
 A \code{logical(1)} determining whether to plot the input error rates (dashed).}

\item{nominal_q}{Optional. Default FALSE.
 A \code{logical(1)} determining whether to plot the expected error rate (red) if the
 quality score exactly matched its nominal definition: Q = -10 log10(p_err).}

\item{...}{Further arguments that are passed to the ggplot() function.}
}
\value{
A \code{\link{ggplot2}} object that will be rendered
 to default device if \code{\link{print}ed},
 or can be stored and further modified.
}
\description{
This function plots the observed freuency of substitutions for each transition
(eg. A->C) as a function of the associated quality score. It also plots the error
rates passed in to this instantiation of the dada algorithm, and the output error
rates if they exist.
}

