% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plotQualityProfile}
\alias{plotQualityProfile}
\title{Plot quality profile of a fastq file.}
\usage{
plotQualityProfile(fl)
}
\arguments{
\item{fl}{(Required). \code{character(1)}.
The file path to the fastq or fastq.gz file.}
}
\value{
A \code{\link{ggplot}} object that will be rendered
 to default device if \code{\link{print}ed},
 or can be stored and further modified.
}
\description{
This function plots a visual summary of the distribution of quality scores
along the sequences within a fastq file. This function returns a \link{ggplot}
graphics object, which can be further modified in ggplot2 fashion.
}
\examples{
plotQualityProfile(system.file("extdata", "sam1F.fastq.gz", package="dada2"))

}

