% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plotQualityProfile}
\alias{plotQualityProfile}
\title{Plot quality profile of a fastq file.}
\usage{
plotQualityProfile(fl, n = 1e+06)
}
\arguments{
\item{fl}{(Required). \code{character(1)}.
The file path to the fastq or fastq.gz file.}

\item{n}{(Optional). Default 1,000,000.
The number of records to sample from the fastq file.}
}
\value{
A \code{\link{ggplot}2} object.
 Will be rendered to default device if \code{\link{print}ed},
 or can be stored and further modified.
 See \code{\link{ggsave}} for additional options.
}
\description{
This function plots a visual summary of the distribution of quality scores
as a function of sequence position for the input fastq file.
}
\examples{
plotQualityProfile(system.file("extdata", "sam1F.fastq.gz", package="dada2"))

}

