% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{qtables2}
\alias{qtables2}
\title{Internal tables function}
\usage{
qtables2(x, qeff = FALSE)
}
\arguments{
\item{x}{ShortReadQ.
The ShortReadQ-class object to table (or dereplicate).}

\item{qeff}{\code{logical(1)}.
Calculate average quality by first transforming to expected error rate.}
}
\value{
List.
 Matches format of derep-class object.
}
\description{
Internal function to replicate ShortRead::tables functionality while also returning average quals and a map
from reads to uniques
}
\keyword{internal}
