% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{seqComplexity}
\alias{seqComplexity}
\title{Determine if input sequence(s) are low complexity.}
\usage{
seqComplexity(seqs, wordSize = 2, window = NULL, by = 5)
}
\arguments{
\item{seqs}{(Required). A \code{character} vector of A/C/G/T sequences, or
any object coercible by \code{\link{getSequences}}.}

\item{wordSize}{(Optional). Default 2.
The size of the oligonucleotides (or "words" or "kmers") to use.}

\item{window}{(Optional). Default NULL.
The width in nucleotides of the moving window. If NULL the whole sequence is used.}

\item{by}{(Optional). Default 5.
The step size in nucleotides between each moving window tested.}
}
\value{
\code{numeric}.
 A vector of minimum olignucleotide complexities for each sequence.
}
\description{
This function calculates the oligonucleotide
complexity of input sequences. Complexity is quantified as the Shannon
richness of oligonucleotides, which can be thought of as the
effective number of oligonucleotides if they were all
at equal frequencies. If a window size is provided, the minimum Shannon
richness observed over sliding window along the sequence is returned.
}
\details{
This function can be used to identify potentially artefactual or undesirable
low-complexity sequences, or sequences with low-complexity regions, as are
sometimes observed in Illumina sequencing runs. When such artefactual
sequences are present, a simple plot of the Shannon oligonucleotide
richness values returned by this function will typically show a clear
bimodal signal.
}
\examples{
sq.norm <- "TACGGAAGGTCCGGGCGTTATCCGGATTTATTGGGTTTAAAGGGAGCGTAGGCCGGAGATTAAGCGTGTTGTGA"
sq.lowc <- "TCCTTCTTCTCCTCTCTTTCTCCTTCTTTCTTTTTTTTCCCTTTCTCTTCTTCTTTTTCTTCCTTCCTTTTTTC"
sq.part <- "TTTTTCTTCTCCCCCTTCCCCTTTCCTTTTCTCCTTTTTTCCTTTAGTGCAGTTGAGGCAGGCGGAATTCGTGG"
sqs <- c(sq.norm, sq.lowc, sq.part)
seqComplexity(sqs)
seqComplexity(sqs, window=25)

}
\seealso{
\code{\link[Biostrings]{oligonucleotideFrequency}}
}
