% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dadaOptions.R
\name{setDadaOpt}
\alias{setDadaOpt}
\title{Set DADA options}
\usage{
setDadaOpt(option, value)
}
\arguments{
\item{option}{(Required). Character.
 The DADA options to set.}

\item{value}{(Required). Corresponding variables of the appropriate data type.
 option and value must have the same length.
 WARNING: DATA TYPE CONSISTENCY IS NOT CURRENTLY CHECKED!!}
}
\description{
Set DADA options
}
\seealso{
\code{\link{getDadaOpt}}
}

