% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show,derep-method}
\alias{show,derep-method}
\alias{show,dada-method}
\title{method extensions to show for dada2 objects.}
\usage{
\S4method{show}{derep}(object)

\S4method{show}{dada}(object)
}
\arguments{
\item{object}{Any R object}
}
\value{
NULL.
}
\description{
See the general documentation of \code{\link[methods]{show}} method for
expected behavior.
}
\seealso{
\code{\link[methods]{show}}
}
