% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,derep-method}
\alias{show,dada-method}
\alias{show,derep-method}
\title{method extensions to show for dada2 objects.}
\usage{
\S4method{show}{derep}(object)

\S4method{show}{dada}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
See the general documentation of \code{\link[methods]{show}} method for
expected behavior.
}
\examples{
# examples
}
\seealso{
\code{\link[methods]{show}}
}

