% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{uniquesToFasta}
\alias{uniquesToFasta}
\title{Write a uniques vector to a FASTA file}
\usage{
uniquesToFasta(unqs, fout, mode = "w")
}
\description{
Basically a wrapper for writeFastq in the ShortRead package. Format is suitable for uchime.
}

