% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceIO.R
\name{uniquesToFasta}
\alias{uniquesToFasta}
\title{Write a uniques vector to a FASTA file}
\usage{
uniquesToFasta(unqs, fout, ids = NULL, mode = "w", width = 20000, ...)
}
\arguments{
\item{unqs}{The uniques vector. E.g. the $uniques from the output of \code{\link{derepFastq}}.}

\item{fout}{The file path of the output file. The file you want to write.}

\item{ids}{A character vector of sequence ids, one for each element in \code{unqs}.
Default value is \code{NULL}, in which case a uchime-compatible ID is assigned.}

\item{mode}{A character string flag passed on to \code{\link[ShortRead]{writeFasta}}
indicating the type of file writing mode. Default is \code{"w"}.}

\item{width}{The number of characters per line in the file. Default is 20000.
Passed on to \code{\link[ShortRead]{writeFasta}}.}

\item{...}{Additional parameters passed on to \code{\link[ShortRead]{writeFasta}}.}
}
\value{
NULL.
}
\description{
A wrapper for writeFastq in the ShortRead package.
Default output format is compatible with uchime.
}
\examples{
derep1 = derepFastq(system.file("extdata", "sam1F.fastq.gz", package="dada2"))
outfile <- tempfile(fileext=".fasta")
uniquesToFasta(getUniques(derep1), outfile)
uniquesToFasta(getUniques(derep1), outfile, ids=paste0("Sequence", seq(length(getUniques(derep1)))))

}

